requirements.txt
arrow==1.3.0
asgiref==3.7.2
binaryornot==0.4.4
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.5.0
cssbeautifier==1.14.11
Django==5.0.1
django-browser-reload==1.12.1
django-ckeditor==6.7.1
django-dotenv==1.4.2
django-js-asset==2.2.0
django-tailwind==3.8.0
django-tinymce==3.6.1
djlint==1.34.1
EditorConfig==0.12.3
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.6
Jinja2==3.1.3
jsbeautifier==1.14.11
json5==0.9.14
markdown-it-py==3.0.0
MarkupSafe==2.1.4
mdurl==0.1.2
pathspec==0.12.1
pillow==10.2.0
psycopg2-binary==2.9.9
Pygments==2.17.2
python-dateutil==2.8.2
python-slugify==8.0.2
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tqdm==4.66.1
types-python-dateutil==2.8.19.20240106
urllib3==2.1.0
